Controller and templates update

+++ src/main/java/net/henryco/hblog/mvc/servives/HomePageService.java
@@ -23,7 +23,7 @@ public class HomePageService {
 
 
 
-	public List<StandardPost> getLastPosts(long numb) {
+	public List<StandardPost> getLastPosts(int numb) {
 		return postFormDao.getLastPosts(numb);
 	}
